% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simple}
\alias{simple}
\title{Simple Function}
\usage{
simple(x, y)
}
\arguments{
\item{x}{value}

\item{y}{value}
}
\value{
array of x,y
}
\description{
Simple Function
}
\examples{
x <- 1
y <- 2
simple(x, y)

}
